Working Diagonal tree blocks specially targeting palm trees only
