--- src/main/java/net/tropicraft/core/mixins/AbstractBlockstateAccessor.java
@@ -1,13 +1,0 @@
-package net.tropicraft.core.mixins;
-
-import net.minecraft.block.AbstractBlock;
-import net.minecraft.block.BlockState;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.gen.Invoker;
-
-
-@Mixin(AbstractBlock.AbstractBlockState.class)
-public interface AbstractBlockstateAccessor {
-    @Invoker("asBlockState")
-        BlockState invokeAsBlockState();
-}
